deleteRedditor endpoint

+++ src/main/java/danielrichtersz/controllers/impl/RedditorControllerImpl.java
@@ -10,6 +10,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import javax.xml.ws.Response;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -57,26 +58,51 @@ public class RedditorControllerImpl implements RedditorController {
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body("The given username did not match an existing user's username");
     }
 
+    @PutMapping("/redditors")
+    @ApiOperation(value = "Edit a specific redditor by ID", response = Redditor.class)
+    @Override
+    public ResponseEntity editRedditor(
+            @ApiParam(value = "The id for the to-be edited redditor", required = true) @RequestParam(value = "id") Long id,
+            @ApiParam(value = "The old password of the redditor, for verification") @RequestParam(value = "oldpassword") String oldpassword,
+            @ApiParam(value = "The new password of the redditor") @RequestParam(value = "newpassword") String newpassword) {
+        Redditor redditor = redditorService.findById(id);
 
+        if (redditor == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Redditor not found");
+        }
 
-    /*@PutMapping("/redditors/{id}")
-    @ApiOperation(value = "Update a redditor by id")
-    public ResponseEntity<Redditor> updateRedditor(@ApiParam(value = "The id of the to-be updated redditor", required = true)
-                                                       @PathVariable(value = "id") Long userID,
-                                                   @ApiParam(value = "The new username of the to-be updated redditor", required = true)
-                                                        @PathVariable(value = "name") String username) {
+        if (!redditor.passwordIsValid(oldpassword)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("The provided current password did not match the password for this user");
+        }
 
+        redditor.setPassword(newpassword);
 
-    }*/
+        Redditor editedRedditor = redditorService.editRedditor(redditor);
 
-    @DeleteMapping("/redditors/{id}")
+        if (editedRedditor == null) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Something went wrong, please try again or contact customer support");
+        }
+
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body(editedRedditor);
+    }
+
+    @DeleteMapping("/redditors")
     @ApiOperation(value = "Delete a redditor by id")
     @Override
-    public Map<String, Boolean> deleteRedditor(@ApiParam(value = "The id of the to-be deleted redditor", required = true)
-                                               @PathVariable(value = "id") Long userID) {
+    public ResponseEntity deleteRedditor(@ApiParam(value = "The id of the to-be deleted redditor", required = true)
+                                         @RequestParam(value = "id") Long userID,
+                                         @ApiParam(value = "The password of the to-be deleted redditor", required = true)
+                                         @RequestParam(value = "password") String password) {
+        Redditor redditor = redditorService.findById(userID);
+        if (redditor == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Redditor not found");
+        }
+
+        if (!redditor.passwordIsValid(password)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Given password did not match the password for this Redditor");
+        }
+
         redditorService.deleteRedditor(userID);
-        Map<String, Boolean> response = new HashMap<>();
-        response.put("deleted", Boolean.TRUE);
-        return response;
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body("Redditor succesfully deleted");
     }
 }
